test find by id arreglado

+++ src/main/java/com/ecomarket/registro_usuario/service/UsuarioService.java
@@ -13,7 +13,7 @@ public class UsuarioService {
     public Usuario crear_usuario(Usuario usuario) {
         return usuarioRepository.save(usuario);
     }
-    public Usuario findById(int id) {
+    public Usuario findById(long id) {
             return usuarioRepository.findById(id);
        
     }
